update rem function to accept f32 parameters and adjust unit tests accordingly

+++ scalar/src/lib.rs
@@ -14,7 +14,7 @@ pub fn quo(a: f32, b: f32) -> f32 {
     a / b
 }
 
-pub fn rem(a: i32, b: i32) -> i32 {
+pub fn rem(a: f32, b: f32) -> f32 {
     a % b
 }
 
@@ -36,7 +36,7 @@ mod tests {
         let quo_result = quo(2.0, 2.0);
         assert_eq!(quo_result, 1.0);
 
-        let rem_result = rem(2, 2);
-        assert_eq!(rem_result, 0);
+        let rem_result = rem(2.0, 2.0);
+        assert_eq!(rem_result, 0.0);
     }
 }
